feat(models): add workflow run sharing tables (#216)
Adds a table and an enum to the database. The `user_workflow` table
makes it possible to verify whether a specific workflow has been shared
with a user. The `accesstype` enum defines the different access types
that can be granted to a user. For now, only `read` access is supported.

Closes #206