models: improve indexes
Add a new index to `__reana.job` on `workflow_uuid` and `created` to
query efficiently the jobs that are part of a given workflow, with the
possibility of ordering them by creation time.

Add a new index to `__reana.workflow` on `status` to efficiently query
the total number of running workflows.

Change the already-existing index on `__reana.workflow` so that
`owner_id` is the leading column, before `name`.

Closes #211