fix(models): remove duplicated User-Workflow relationship (#234)
Fix SQLAlchemy warning reporting conflicting relationships:

```
SAWarning: relationship 'User.workflow' will copy column user_.id_ to
column workflow.owner_id, which conflicts with relationship(s):
'User.workflows' (copies user_.id_ to workflow.owner_id),
'Workflow.user_' (copies user_.id_ to workflow.owner_id). If this is not
the intention, consider if these relationships should be linked with
back_populates, or if viewonly=True should be applied to one or more if
they are read-only. For the less common case that foreign key
constraints are partially overlapping, the orm.foreign() annotation can
be used to isolate the columns that should be written towards.   To
silence this warning, add the parameter 'overlaps="user_,workflows"' to
the 'User.workflow' relationship. (Background on this error at:
https://sqlalche.me/e/14/qzyx)
```